+++ src/lib/types/index.ts
@@ -13,6 +13,7 @@ export * from "./converter";
 export * from "./delegation";
 export * from "./json";
 export * from "./LVPair";
+export * from "./nft";
 export * from "./pool";
 export * from "./projects";
 export * from "./proposal";