+++ src/lib/types/tx/transaction.ts
@@ -84,3 +84,9 @@ export type PoolTxFilter =
 
 export const zRemarkType = z.enum(["genesis", "governance", "transaction"]);
 export type RemarkType = z.infer<typeof zRemarkType>;
+
+export const zRemark = z.object({
+  type: zRemarkType,
+  value: z.union([z.string(), z.number()]).optional(),
+});
+export type Remark = z.infer<typeof zRemark>;