+++ src/pages/[network]/nft-collections/[collectionAddress]/nft/[nftAddress].tsx
@@ -1,0 +1,3 @@
+import NftDetails from "lib/pages/nft-details";
+
+export default NftDetails;